+++ CHANGELOG.md
@@ -4,6 +4,7 @@ AppEngine version, listed here to ease deployment and troubleshooting.
 ## Next Release (replace with git tag when deployed)
  * Bumped runtimeVersion to `2022.07.07`.
  * Upgraded dartdoc to `6.0.0`.
+ * NOTE: Require a single LICENSE file (any variant), reject if there are multiple ones.
 
 ## `20220705t102600-all`
  * Bumped runtimeVersion to `2022.07.05`.